Add correct_divE to cudametods, fix sort array dtype

+++ fbpic/particles/particles.py
@@ -215,8 +215,8 @@ class Particles(object) :
             # Allocate arrays for the particles sorting when using CUDA
             # Most required arrays always stay on GPU
             Nz, Nr = grid_shape
-            self.cell_idx = cuda.device_array( Ntot, dtype=np.int32)
-            self.sorted_idx = cuda.device_array( Ntot, dtype=np.intp)
+            self.cell_idx = cuda.device_array( Ntot, dtype=np.int64)
+            self.sorted_idx = cuda.device_array( Ntot, dtype=self.cell_idx.dtype )
             self.prefix_sum = cuda.device_array( Nz*(Nr+1), dtype=np.int32 )
             # sorting buffers are initialized on CPU like other particle arrays
             # (because they are swapped with these arrays during sorting)
